Add Ninji Wallet (#37)
Add Ninji Wallet Extension

Ninji is your agile Injective Wallet for a fast and effortless
experience in managing assets and connecting to Injective dApps.

Ninji is designed for passionate users seeking an efficient and
straightforward solution for managing their digital assets and access to
the truly decentralization. By offering ease, security, and instant
connectivity, Ninji unlocks the full potential for a broad audience,
setting itself apart as the go-to wallet for effortless DeFi
experiences. Its exceptional offerings include:

+ Seamlessly send & receive assets
+ Solely own your self-custodial wallet
+ Effortlessly secure your privacy & security
+ Freely explore DeFi with a friendly user interface
+ Instantly connect to Injective dApps
+ Personalize your interface with customized light and dark modes.

Install now to experience seamless Decentralize Finance with Ninji.

We are always here to help:
+ Twitter: https://twitter.com/ninjiwallet
+ Discord: https://t.co/YXndeEhNII